Midway work for converting to postgresql
Converting original proejct named "TweetScanner" from couchdb to use
postgres. This project still uses Java8 and Redis which really is the
heart of the application anyway. Also added Vagrant and ansible to make
it very easy to build the application locally. It was already easy but
it is nice to be able to just pick up the project and having it working
anytime, plus any dev-environment automation pays off very very quickly.