Using Project Reactor to orchestrate processing
Able to subscribe to tweets as they come in, on multiple subscribers.
Will re-add remaining processing (word counts, persistence).

Also implemented the first iteration of the tweet spam filter. Will
change it to use redis keys split into buckets (presharding), as it is
not possible to put a TTL expiry on elements in a set. And having a key
for every tweet has a major memory consumption impact.

Need to remove the flush counter, flushing occurs now when a tweet is
saved for spam prevention.